Split controllers and models
