+++ server/controllers/index.js
@@ -1,7 +1,18 @@
-//import express from 'express';
-//var router = express.Router();
-//
-//var Utils      = require(path.join(__dirname, './utils.js'));
-//var db         = require(path.join(__dirname, './db.js'));
-//
-//export default router;
+var express = require('express');
+var router = express.Router();
+var Art = require('../models/art');
+
+module.exports = router;
+
+router.use('/auth', require('./auth'));
+router.use('/favorite', require('./favorite'));
+router.use('/like', require('./like'));
+
+// client asking for art data
+router.get('/art', function(req, res) {
+  //retrieve all art from db
+  Art.get()
+    .then(art => {
+      res.send(art);
+    })
+})
