+++ server/models/art.js
@@ -1,0 +1,6 @@
+var db = require('../db');
+
+
+exports.get = () => {
+  return db.art.find();
+};
